PS-5 Added error handling to user meal plan service

+++ services/userMealPlan.service.js
@@ -1,21 +1,42 @@
 const MealPlan = require("../models/meal-plan.model");
+const { NotFoundError } = require("../helpers/errorHandlers");
 
 const getUserMealPlans = async (user_id) => {
-  return await MealPlan.find({ author: user_id });
+  const mealPlan = await MealPlan.find({ author: user_id });
+  if (!mealPlan) {
+    throw new NotFoundError("Meal plans don't exist");
+  } else {
+    return mealPlan;
+  }
 };
 
 const getUserMealPlan = async (id) => {
-  return await MealPlan.findById({ _id: id });
+  const mealPlan = await MealPlan.findById({ _id: id });
+  if (!mealPlan) {
+    throw new NotFoundError("Meal plan doesn't exist");
+  } else {
+    return mealPlan;
+  }
 };
 
 const udpateUserMealPlan = async (id, requestBody) => {
-  return await MealPlan.findByIdAndUpdate({ _id: id }, requestBody, {
-    new: true,
-  });
+  const mealPlan = await MealPlan.findById({ _id: id });
+  if (!mealPlan) {
+    throw new NotFoundError("Meal plan doesn't exist");
+  } else {
+    return await MealPlan.findByIdAndUpdate({ _id: id }, requestBody, {
+      new: true,
+    });
+  }
 };
 
 const deleteUserMealPlan = async (id) => {
-  return await MealPlan.deleteOne({ _id: id });
+  const mealPlan = await MealPlan.findById({ _id: id });
+  if (!mealPlan) {
+    throw new NotFoundError("Meal plan doesn't exist");
+  } else {
+    return await MealPlan.deleteOne({ _id: id });
+  }
 };
 
 const createUserMealPlan = async (request) => {
